// overlay/ui/src/generated/proto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BorderStyle = "off" | "faint" | "medium" | "strong";

export type PixelRect = { x: number, y: number, w: number, h: number, };

export type Size = { w: number, h: number, };

export type SlotState = { rect: PixelRect, live: boolean, status: SlotStatus | null, label: string | null, avatar: string | null, logo: string | null, controller_disconnected: boolean, };

export type SlotStatus = "loading" | "running";

export type State = { proto: number, size: Size, focus: number, border: BorderStyle, slots: Array<SlotState>, };